data/pyproject.toml
[project]
name = "build"
requires-python = ">=3.11"
dependencies = [
    "freetype-py",
    "numpy",
]
